Add new suggestion for Case Study
Recognition memory experiments give us speed and accuracy data. Both of
this could be informative source of information about memory, even
though the usual procedure in memory researc is to only analyse accuracy.
To analyse both accuracy and reaction time, the ideal method is to use a
sequential sampling model like the drift diffusion model. This is what
will be showcased here.