Add script for downloading data
The data for one case study, i.e., the study on reaction times and
accuracies in recognition memory experiments is from

  Kim, S., Potter, K., Craigmile, P. F., Peruggia, M., & Van Zandt, T. (2017). A
  Bayesian Race Model for Recognition Memory. Journal of the American Statistical
  Association, 112(517), 77-91.

and is available from GitHub.

The shell script `get_data.sh` gets this data. The `md5sum_check.txt` is
used to check if the data files are correct.